task done

+++ contest/codechef/20th-march-div-4/Sale.cpp
@@ -26,7 +26,7 @@ int main()
                 ind = i;
             }
         }
-        int ans = 0;
+        ll ans = 0;
         for (int i = 0; i < n; i++)
         {
             if (i == ind)
